Ported to version 2.6.37 due to ARM compatibilities

+++ borph/arch/ia64/kernel/efi.c
@@ -23,7 +23,6 @@
  */
 #include <linux/module.h>
 #include <linux/bootmem.h>
-#include <linux/crash_dump.h>
 #include <linux/kernel.h>
 #include <linux/init.h>
 #include <linux/types.h>
@@ -156,7 +155,7 @@ prefix##_get_next_variable (unsigned long *name_size, efi_char16_t *name,      \
 #define STUB_SET_VARIABLE(prefix, adjust_arg)				       \
 static efi_status_t							       \
 prefix##_set_variable (efi_char16_t *name, efi_guid_t *vendor,		       \
-		       u32 attr, unsigned long data_size,		       \
+		       unsigned long attr, unsigned long data_size,	       \
 		       void *data)					       \
 {									       \
 	struct ia64_fpreg fr[6];					       \
